+++ DependencyInjection/Ma27ApiKeyAuthenticationExtension.php
@@ -50,11 +50,7 @@ class Ma27ApiKeyAuthenticationExtension extends Extension
      */
     private function loadPassword(ContainerBuilder $container, $passwordConfig, Loader\YamlFileLoader $loader)
     {
-        // TODO refactor this, the `phpass_iteration_length` is specific for a single strategy and therefore it shouldn't be present in the semantic configuration.
-        $container->setParameter(
-            'ma27_api_key_authentication.password_hasher.phpass.iteration_length',
-            isset($passwordConfig['phpass_iteration_length']) ? $passwordConfig['phpass_iteration_length'] : 8
-        );
+        $container->setParameter('ma27_api_key_authentication.password_hasher.phpass.iteration_length', 8);
         $loader->load('hashers.yml');
 
         $container->setParameter(